+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### Version 0.0.4
+
+  - adds a context object ``this.ctx`` that is share between steps in a scenario
+  - bumps yadda to 0.15.4 which includes a library for dutch
+
+
 ### Version 0.0.3
 
   - fixes a bug with the build order ember-cli-yadda needs to go __before__ babel